LPS-111617 Extract ConfigurationModelListener
